Implement authorization checks in StudentController for various actions

+++ source/app/Http/Controllers/Api/StudentController.php
@@ -2,9 +2,8 @@
 
 namespace App\Http\Controllers\Api;
 
-use App\Enums\LogLevel;
 use App\Http\Controllers\Controller;
-use App\Models\Log;
+use App\Models\Admin;
 use App\Models\Student;
 use App\Services\LogService;
 use App\Services\MailService;
@@ -18,6 +17,7 @@ class StudentController extends Controller
     public function index(): JsonResponse
     {
         $students = Student::all();
+        $this->authorize("viewAny", Admin::class);
 
         return response()->json([
             'data' => $students
@@ -60,9 +60,11 @@ class StudentController extends Controller
         }
 
         $validated['password'] = Hash::make($validated['password']);
+
+        $this->authorize("create", Student::class);
         $student = Student::create($validated);
 
-        $logService->setLog("Student", $student->id,"Student created", "Student");
+        $logService->setLog("Student", $student->id, "Student created", "Student");
         $mailService->sendStudentVerification($student, $logService);
 
         return response()->json([
@@ -79,42 +81,8 @@ class StudentController extends Controller
     {
         try {
             $student = Student::findOrFail($id);
+            $this->authorize("view", $student);
             return response()->json(['data' => $student]);
-        } catch (ModelNotFoundException $e) {
-            return response()->json(['message' => 'Student not found'], 404);
-        }
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(Request $request, string $id, LogService $logService): JsonResponse
-    {
-        try {
-            $student = Student::findOrFail($id);
-            $validated = $request->validate([
-                'first_name' => 'required|string|max:255',
-                'last_name' => 'required|string|max:255',
-                'email' => 'required|email|unique:students,email,' . $student->id,
-                'password' => 'required|string|min:8',
-                'study_direction' => 'required|string|max:255',
-                'graduation_track' => 'required|integer|exists:diplomas,id',
-                'interests' => 'required|string',
-                'job_preferences' => 'required|string',
-                'cv' => 'nullable|string|max:255',
-                'profile_complete' => 'boolean',
-                'profile_photo' => 'nullable|string|max:255',
-            ]);
-
-            $validated['password'] = Hash::make($validated['password']);
-            $student->update($validated);
-
-            $logService->setLog("Student", $student->id, "Student updated", "Student");
-
-            return response()->json([
-                'data' => $student,
-                'message' => 'Student updated successfully'
-            ]);
         } catch (ModelNotFoundException $e) {
             return response()->json(['message' => 'Student not found'], 404);
         }
@@ -124,6 +92,8 @@ class StudentController extends Controller
     {
         try {
             $student = Student::findOrFail($id);
+
+            $this->authorize("delete", $student);
             $student->delete();
 
             $logService->setLog("Student", $student->id, "Student deleted", "Student");
@@ -136,11 +106,11 @@ class StudentController extends Controller
         }
     }
 
-
     public function verify(string $id)
     {
         try {
             $student = Student::findOrFail($id);
+            $this->authorize("verify", $student);
             if ($student->profile_complete) {
                 return response()->json(['message' => 'Student was already verified']);
             } else {
@@ -180,14 +150,52 @@ class StudentController extends Controller
                 $data['password'] = Hash::make($data['password']);
             }
 
+            $this->authorize("update", $student);
             $student->update($data);
-            $logService->setLog("Student", $student->id,"Student updated", "Student");
+            $logService->setLog("Student", $student->id, "Student updated", "Student");
 
             return response()->json([
                 'data' => $student,
                 'message' => 'Student partially updated successfully',
             ]);
 
+        } catch (ModelNotFoundException $e) {
+            return response()->json(['message' => 'Student not found'], 404);
+        }
+    }
+
+    /**
+     * Update the specified resource in storage.
+     */
+    public function update(Request $request, string $id, LogService $logService): JsonResponse
+    {
+        try {
+            $student = Student::findOrFail($id);
+            $validated = $request->validate([
+                'first_name' => 'required|string|max:255',
+                'last_name' => 'required|string|max:255',
+                'email' => 'required|email|unique:students,email,' . $student->id,
+                'password' => 'required|string|min:8',
+                'study_direction' => 'required|string|max:255',
+                'graduation_track' => 'required|integer|exists:diplomas,id',
+                'interests' => 'required|string',
+                'job_preferences' => 'required|string',
+                'cv' => 'nullable|string|max:255',
+                'profile_complete' => 'boolean',
+                'profile_photo' => 'nullable|string|max:255',
+            ]);
+
+            $validated['password'] = Hash::make($validated['password']);
+
+            $this->authorize("update", $student);
+            $student->update($validated);
+
+            $logService->setLog("Student", $student->id, "Student updated", "Student");
+
+            return response()->json([
+                'data' => $student,
+                'message' => 'Student updated successfully'
+            ]);
         } catch (ModelNotFoundException $e) {
             return response()->json(['message' => 'Student not found'], 404);
         }
